refactor(ws-6): split executor.rs into topic submodules
